Got rid of the manual entry areas
Eliminated the manual Entry areas, rn it's out of our scope. But in the future we will create one that will both randomly generate an NPC or give the viewer the option to manually enter NPC info, randomly generate empty fields, allow them to save those NPCs as text files, and also add manually entered info in to the array lists for random generation.